Scripts for the simulation of immune evolution
The Julia programming language is necessary to recreate this work, if starting from scratch please follow the instructions here to begin (https://github.com/julia-vscode/julia-vscode#installing-juliavs-codevs-code-julia-extension )

The above scripts are those necessary to recreate my simulations of immune evolution under the influence of pleiotropic signaling proteins. Each script that starts with "Simulation" outputs a data file that can then be read by an accompanying figure script to produce the specified figure for example to create a Figure 2, you would run each of the "Simulation of Immune NetworkEvo" scripts, and then run the script "Final_f2" to create that figure.
If you have any comments or questions about this work please send me a message and I will get back to you as time allows.